+++ app/src/main/java/com/zhiyicx/thinksnsplus/modules/dynamic/list/DynamicPresenter.java
@@ -297,6 +297,7 @@ public class DynamicPresenter extends AppBasePresenter<DynamicContract.Repositor
     @Override
     public void reSendDynamic(int position) {
         // 将动态信息存入数据库
+
         mDynamicDetailBeanV2GreenDao.insertOrReplace(mRootView.getListDatas().get(position));
         // 发送动态
         BackgroundRequestTaskBean backgroundRequestTaskBean = new BackgroundRequestTaskBean();
@@ -340,6 +341,9 @@ public class DynamicPresenter extends AppBasePresenter<DynamicContract.Repositor
 
     @Override
     public void deleteDynamic(DynamicDetailBeanV2 dynamicBean, int position) {
+        if (position == -1) {
+            return;
+        }
         mDynamicDetailBeanV2GreenDao.deleteSingleCache(dynamicBean);
         mRootView.getListDatas().remove(position);
         mRootView.refreshData();
